+++ sentry/src/test/java/io/sentry/metrics/MetricsApiTest.kt
@@ -1,13 +1,17 @@
 package io.sentry.metrics
 
 import io.sentry.IMetricsAggregator
+import io.sentry.ISpan
+import io.sentry.MeasurementUnit
 import io.sentry.metrics.MetricsApi.IMetricsInterface
 import org.mockito.kotlin.any
 import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
 import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class MetricsApiTest {
 
@@ -15,7 +19,14 @@ class MetricsApiTest {
         val aggregator = mock<IMetricsAggregator>()
         val localMetricsAggregator = mock<LocalMetricsAggregator>()
 
-        fun getSut(defaultTags: Map<String, String> = emptyMap()): MetricsApi {
+        var lastSpan: ISpan? = null
+        var lastOp: String? = null
+        var lastDescription: String? = null
+
+        fun getSut(
+            defaultTags: Map<String, String> = emptyMap(),
+            spanProvider: () -> ISpan? = { mock<ISpan>() }
+        ): MetricsApi {
             val localAggregator = localMetricsAggregator
 
             return MetricsApi(object : IMetricsInterface {
@@ -26,6 +37,13 @@ class MetricsApiTest {
                 override fun getLocalMetricsAggregator(): LocalMetricsAggregator? = localAggregator
 
                 override fun getDefaultTagsForMetrics(): Map<String, String> = defaultTags
+
+                override fun startSpanForMetric(op: String, description: String): ISpan? {
+                    lastOp = op
+                    lastDescription = description
+                    lastSpan = spanProvider()
+                    return lastSpan
+                }
             })
         }
     }
@@ -287,5 +305,68 @@ class MetricsApiTest {
             anyOrNull(),
             eq(fixture.localMetricsAggregator)
         )
+    }
+
+    @Test
+    fun `timing starts and finishes a span`() {
+        val api = fixture.getSut()
+
+        api.timing("key") {
+            // no-op
+        }
+
+        assertEquals("metric.timing", fixture.lastOp)
+        assertEquals("key", fixture.lastDescription)
+
+        verify(fixture.lastSpan!!).finish()
+    }
+
+    @Test
+    fun `timing applies metric tags as span tags`() {
+        val span = mock<ISpan>()
+        val api = fixture.getSut(
+            spanProvider = {
+                span
+            },
+            defaultTags = mapOf(
+                "release" to "1.0"
+            )
+        )
+        // when timing is called
+        api.timing("key", {
+            // no-op
+        }, MeasurementUnit.Duration.NANOSECOND, mapOf("a" to "b"))
+
+        // the last span should have the metric tags, without the default ones
+        verify(fixture.lastSpan!!, never()).setTag("release", "1.0")
+        verify(fixture.lastSpan!!).setTag("a", "b")
+    }
+
+    @Test
+    fun `if timing throws an exception, span still finishes`() {
+        val api = fixture.getSut()
+
+        try {
+            api.timing("key") {
+                throw IllegalStateException()
+            }
+        } catch (e: IllegalStateException) {
+            // ignored
+        }
+
+        assertEquals("metric.timing", fixture.lastOp)
+        assertEquals("key", fixture.lastDescription)
+        verify(fixture.lastSpan!!).finish()
+    }
+
+    @Test
+    fun `if timing does retrieve a null span, it still works`() {
+        val api = fixture.getSut(
+            spanProvider = { null }
+        )
+        api.timing("key") {
+            // no-op
+        }
+        // no crash
     }
 }